Minor improvements and documentation
